+++ lib/src/shared/message.dart
@@ -16,6 +16,10 @@
 import 'dart:async';
 import 'dart:typed_data';
 
+import 'codec.dart';
+import 'codec_registry.dart';
+import 'status.dart';
+
 abstract class GrpcMessage {}
 
 class GrpcMetadata extends GrpcMessage {
@@ -54,25 +58,38 @@ class GrpcMessageSink extends Sink<GrpcMessage> {
   }
 }
 
-List<int> frame(List<int> payload) {
-  final payloadLength = payload.length;
+List<int> frame(List<int> rawPayload, [Codec codec]) {
+  final compressedPayload =
+      codec == null ? rawPayload : codec.compress(rawPayload);
+  final payloadLength = compressedPayload.length;
   final bytes = Uint8List(payloadLength + 5);
   final header = bytes.buffer.asByteData(0, 5);
-  header.setUint8(0, 0); // TODO(dart-lang/grpc-dart#6): Handle compression
+  header.setUint8(0, codec == null ? 0 : 1);
   header.setUint32(1, payloadLength);
-  bytes.setRange(5, bytes.length, payload);
+  bytes.setRange(5, bytes.length, compressedPayload);
   return bytes;
 }
 
-StreamTransformer<GrpcMessage, GrpcMessage> grpcDecompressor() =>
-    StreamTransformer<GrpcMessage, GrpcMessage>.fromHandlers(
-        handleData: (GrpcMessage value, EventSink<GrpcMessage> sink) {
-      if (value is GrpcData) {
-        if (value.isCompressed) {
-          // TODO(dart-lang/grpc-dart#6): Actually handle decompression.
-          sink.add(GrpcData(value.data, isCompressed: false));
-          return;
-        }
+StreamTransformer<GrpcMessage, GrpcMessage> grpcDecompressor({
+  CodecRegistry codecRegistry,
+}) {
+  Codec codec;
+  return StreamTransformer<GrpcMessage, GrpcMessage>.fromHandlers(
+      handleData: (GrpcMessage value, EventSink<GrpcMessage> sink) {
+    if (value is GrpcData && value.isCompressed) {
+      if (codec == null) {
+        sink.addError(
+          GrpcError.unimplemented('Compression mechanism not supported'),
+        );
+        return;
       }
-      sink.add(value);
-    });
+      sink.add(GrpcData(codec.decompress(value.data), isCompressed: false));
+      return;
+    }
+
+    if (value is GrpcMetadata && value.metadata.containsKey('grpc-encoding')) {
+      codec = codecRegistry?.lookup(value.metadata['grpc-encoding']);
+    }
+    sink.add(value);
+  });
+}